+++ Editor/Content/DefaultAssets.cs
@@ -1,8 +1,10 @@
 ﻿using Editor.Common.Enums;
 using Editor.DLLs;
 using Editor.DLLs.Descriptors;
+using Editor.Utilities;
 using System.Diagnostics;
 using System.IO;
+using System.Windows;
 
 namespace Editor.Content
 {
@@ -10,6 +12,7 @@ namespace Editor.Content
     {
         public static AssetInfo BRDFIntegrationLUT { get; private set; }
         public static AssetInfo DefaultGeometry {  get; private set; }
+        public static AssetInfo DefaultMaterial { get; private set; }
 
         /// <summary>
         /// Generate default assets if necessary
@@ -27,6 +30,14 @@ namespace Editor.Content
             var cubeFileName = $@"{defaultAssetsPath}DefaultCube{Asset.AssetFileExtension}";
 
             if (!File.Exists(cubeFileName)) CreateDefaultCube(cubeFileName);
+
+            var materialFileName = $@"{defaultAssetsPath}DefaultMaterial{Asset.AssetFileExtension}";
+
+            if (!File.Exists(materialFileName)) CreateDefaultMaterial(materialFileName);
+
+            BRDFIntegrationLUT = Asset.GetAssetInfo(brdfLUTFileName);
+            DefaultGeometry = Asset.GetAssetInfo(cubeFileName);
+            //DefaultMaterial = Asset.GetAssetInfo(materialFileName);
         }
 
         private static void ComputeBRDFIntegrationLUT(string file)
@@ -40,8 +51,6 @@ namespace Editor.Content
 
                 ContentToolsAPI.ComputeBRDFIntegrationLUT(brdfLUT);
                 brdfLUT.Save(file);
-
-                BRDFIntegrationLUT = Asset.GetAssetInfo(file);
             }
             catch (Exception ex)
             {
@@ -62,10 +71,89 @@ namespace Editor.Content
 
                 ContentToolsAPI.CreatePrimitiveMesh(cube, info);
                 cube.Save(file);
-
-                DefaultGeometry = Asset.GetAssetInfo(file);
             }
             catch(Exception ex)
+            {
+                Debug.WriteLine(ex.Message);
+            }
+        }
+
+        private static ShaderGroup CompileShaderGroup(
+            ShaderType type,
+            string code,
+            string functionName,
+            string[] defines,
+            uint[] keys
+        )
+        {
+            var extraArgs = new List<List<string>>();
+
+            foreach(var def in defines)
+            {
+                extraArgs.Add(!string.IsNullOrEmpty(def.Trim()) ? new()
+                {
+                    "-D",
+                    def
+                } : new());
+            }
+
+            var shaderGroup = new ShaderGroup()
+            {
+                Type = type,
+                Code = code,
+                FunctionName = functionName,
+                ExtraArgs = extraArgs,
+                Keys = [.. keys]
+            };
+
+            EngineAPI.CompileShader(shaderGroup);
+
+            return shaderGroup;
+        }
+
+        private static void CreateDefaultMaterial(string file)
+        {
+            var vsDefines = new[]
+            {
+                "ELEMENTS_TYPE=0",
+                "ELEMENTS_TYPE=1",
+                "ELEMENTS_TYPE=3"
+            };
+
+            var vsKeys = new[]
+            {
+                (uint)ElementsType.POSITION_ONLY,
+                (uint)ElementsType.STATIC_NORMAL,
+                (uint)ElementsType.STATIC_NORMAL_TEXTURE,
+            };
+
+            var psDefines = new[]
+            {
+                string.Empty,
+            };
+
+            var psKeys = new[]
+            {
+                (uint)Id.InvalidId,
+            };
+
+            try
+            {
+                var code = string.Empty;
+
+                var shaderUri = ContentHelper.GetPackUri(
+                    @"Resources/MaterialEditor/DefaultMaterialShaders.hlsl",
+                    typeof(DefaultAssets)
+                );
+
+                var info = Application.GetResourceStream(shaderUri);
+
+                using (var reader = new StreamReader(info.Stream)) code = reader.ReadToEnd();
+
+                var vertexShaders = CompileShaderGroup(ShaderType.VERTEX, code, "main_vs", vsDefines, vsKeys);
+                var pixelShaders = CompileShaderGroup(ShaderType.PIXEL, code, "main_ps", psDefines, psKeys);
+            }
+            catch (Exception ex)
             {
                 Debug.WriteLine(ex.Message);
             }